+++ src/studioServer.java
@@ -1,6 +1,7 @@
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import sss.idao.DAOFactory;
+import sss.model.Seat;
 import sss.model.Studio;
 
 import javax.servlet.ServletException;
@@ -16,6 +17,7 @@ import java.util.*;
 
 /**
  * Created by zxw on 17-12-18.
+ * Created by lyt
  */
 @WebServlet(name = "studio",urlPatterns = "/api/studio")
 public class studioServer extends HttpServlet {
@@ -25,6 +27,7 @@ public class studioServer extends HttpServlet {
         Writer out = response.getWriter();
         JSONObject json = new JSONObject();
         Studio s = new Studio();
+        Seat seat = new Seat();
         try {
             s.setStudio_name(request.getParameter("studio_name"));
             s.setStudio_row_count(Integer.valueOf(request.getParameter("studio_rows")));
@@ -37,13 +40,15 @@ public class studioServer extends HttpServlet {
             out.write(json.toString());
             return;
         }
-        if (DAOFactory.createStudioDAO().insert(s)){
+        int aa = DAOFactory.createStudioDAO().insert(s);
+        if (aa != -1){
             System.out.println("演出厅插入数据成功");
-
-            System.out.println(s);
             json.put("state",true);
-            System.out.println(s.getStudio_id());
-//            json.put("studio_id",s.getStudio_id());
+            seat.setStudio_id(aa);
+            seat.setSeat_row(s.getStudio_row_count());
+            seat.setSeat_column(s.getStudio_col_count());
+            seat.setSeat_status(0);
+            DAOFactory.createSeatDAO().insert(seat);
             out.write(json.toString());
         }else {
             System.out.println("演出厅插入数据失败");
@@ -149,9 +154,9 @@ public class studioServer extends HttpServlet {
             for (String x : listp) {
                 String z[] = x.split("=");
                 hm.put(z[0], z[1]);
+
             }
         }catch (Exception e){
-
         }
 
 
@@ -165,12 +170,16 @@ public class studioServer extends HttpServlet {
             out.write(json.toString());
             return ;
         }
-        if(DAOFactory.createStudioDAO().delete(id)){
-            json.put("status",true);
-            out.write(json.toString());
-        }else{
+        if (DAOFactory.createSeatDAO().delete(id)){
+            if (DAOFactory.createStudioDAO().delete(id)){
+                json.put("status",true);
+                out.write(json.toString());
+
+            }
+
+        } else{
             json.put("status",false);
-            System.out.print("serr");
+            System.out.print("err");
             out.write(json.toString());
         }
     }
@@ -219,8 +228,8 @@ public class studioServer extends HttpServlet {
             return;
         }
 
-        if(hm.get("studio_detial") != null){
-            stu.setStudio_introduction(hm.get("studio_detial"));
+        if(hm.get("stuio_detial") != null){
+            stu.setStudio_introduction(hm.get("stuio_detial"));
         }
 
         if(DAOFactory.createStudioDAO().update(stu)){
